improved unit tests for SwingUtils is*MouseButton
