+++ src/test/java/VASSAL/tools/swing/SwingUtilsTestIsMouseButtonNonMacOsX.java
@@ -1,45 +1,20 @@
 package VASSAL.tools.swing;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
 import java.awt.event.InputEvent;
 import java.awt.event.MouseEvent;
 import java.util.Arrays;
 import java.util.Collection;
 
 import org.junit.Rule;
-import org.junit.Test;
 import org.junit.contrib.java.lang.system.ProvideSystemProperty;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
 @RunWith(Parameterized.class)
-public class SwingUtilsTestIsMouseButtonNonMacOsX {
-
-  /**
-   * index 0: {@link MouseEvent#getID()}
-   *   e.g. {@link MouseEvent#MOUSE_PRESSED}
-   *
-   * index 1: {@link MouseEvent#getButton()}
-   *   e.g. {@link MouseEvent#BUTTON1}
-   *
-   * index 2: {@link MouseEvent#getModifiersEx()}
-   *   e.g. {@link InputEvent#BUTTON1_DOWN_MASK}
-   */
-  private final int[] mouseEventParams;
-
-  /**
-   * index 0: result for isMouseLeft
-   * index 1: result for isMouseRight
-   */
-  private final boolean[] expectedResult;
+public class SwingUtilsTestIsMouseButtonNonMacOsX extends AbstractSwingUtilsTestIsMouseButton {
 
   public SwingUtilsTestIsMouseButtonNonMacOsX(int[] mouseEventParams, boolean[] expectedResult) {
-    this.mouseEventParams = mouseEventParams;
-    this.expectedResult = expectedResult;
+    super(mouseEventParams, expectedResult);
   }
 
   @Parameterized.Parameters
@@ -91,37 +66,7 @@ public class SwingUtilsTestIsMouseButtonNonMacOsX {
   }
 
   @Rule
-  public final ProvideSystemProperty myPropertyHasMyValue =
-      new ProvideSystemProperty("os.name", "anythingButMacOsx");
-
-  @Test
-  public void testIsLeftMouseButton() {
-    // prepare
-    final MouseEvent mouseEvent = mock(MouseEvent.class);
-    when(mouseEvent.getID()).thenReturn(mouseEventParams[0]);
-    when(mouseEvent.getButton()).thenReturn(mouseEventParams[1]);
-    when(mouseEvent.getModifiersEx()).thenReturn(mouseEventParams[2]);
-
-    // run
-    final boolean result = SwingUtils.isLeftMouseButton(mouseEvent);
-
-    // assert
-    assertThat(result, is(expectedResult[0]));
-  }
-
-  @Test
-  public void testIsRightMouseButton() {
-    // prepare
-    final MouseEvent mouseEvent = mock(MouseEvent.class);
-    when(mouseEvent.getID()).thenReturn(mouseEventParams[0]);
-    when(mouseEvent.getButton()).thenReturn(mouseEventParams[1]);
-    when(mouseEvent.getModifiersEx()).thenReturn(mouseEventParams[2]);
-
-    // run
-    final boolean result = SwingUtils.isRightMouseButton(mouseEvent);
-
-    // assert
-    assertThat(result, is(expectedResult[1]));
-  }
+  public final ProvideSystemProperty osNameProperty =
+    new ProvideSystemProperty("os.name", "anythingButMacOsx");
 
 }
